Reorganizar páginas de autenticación y cuenta
